ref(metrics): Create plot metrics subpackage
Create a subpackage for plot metrics containing separate modules instead of putting everything in the same module. This makes the code better organized.
