+++ docs/source/decent_bench.library.core.metrics.rst
@@ -1,6 +1,14 @@
 decent\_bench.library.core.metrics package
 ==========================================
 
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   decent_bench.library.core.metrics.plot_metrics
+
 Submodules
 ----------
 
@@ -8,14 +16,6 @@ decent\_bench.library.core.metrics.metric\_utils module
 -------------------------------------------------------
 
 .. automodule:: decent_bench.library.core.metrics.metric_utils
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.library.core.metrics.plot\_metrics module
--------------------------------------------------------
-
-.. automodule:: decent_bench.library.core.metrics.plot_metrics
    :members:
    :show-inheritance:
    :undoc-members:
